refactor(ms2): same layout for all auth routes
