+++ src/management-system-v2/app/(auth)/layout.tsx
@@ -1,7 +1,74 @@
+import Layout from '@/app/(dashboard)/[environmentId]/layout-client';
+import Content from '@/components/content';
+import Processes from '@/components/processes';
+import { SetAbility } from '@/lib/abilityStore';
 import { FC, PropsWithChildren } from 'react';
+import { AiOutlineFile, AiOutlineProfile } from 'react-icons/ai';
 
-const AuthLayout: FC<PropsWithChildren> = ({ children }) => {
-  return children;
+const SigninLayout: FC<PropsWithChildren> = ({ children }) => {
+  return (
+    <>
+      {children}
+      <SetAbility rules={[]} environmentId={''} />
+      <Layout
+        activeSpace={{ spaceId: '', isOrganization: false }}
+        loggedIn={true}
+        userEnvironments={[{ isOrganization: false, ownerId: '', id: '' }]}
+        layoutMenuItems={[
+          {
+            key: 'processes-group',
+            label: 'Processes',
+            children: [
+              {
+                key: 'processes',
+                label: 'Process List',
+                icon: <AiOutlineFile />,
+              },
+              {
+                key: 'templates',
+                label: 'Templates',
+                icon: <AiOutlineProfile />,
+              },
+            ],
+          },
+        ]}
+      >
+        <Content title="Processes">
+          <Processes
+            processes={[
+              {
+                id: '',
+                folderId: '',
+                environmentId: '',
+                creatorId: '',
+                name: 'How to PROCEED',
+                type: 'process',
+                versions: [],
+                createdOn: new Date(),
+                lastEditedOn: new Date(),
+                //variables: [],
+                //departments: [],
+                processIds: [],
+                description: 'How to PROCEED',
+                sharedAs: 'protected',
+                shareTimestamp: 0,
+                allowIframeTimestamp: 0,
+              },
+            ]}
+            folder={{
+              id: '',
+              name: '',
+              parentId: null,
+              createdOn: new Date(),
+              createdBy: '',
+              lastEditedOn: new Date(),
+              environmentId: '',
+            }}
+          />
+        </Content>
+      </Layout>
+    </>
+  );
 };
 
-export default AuthLayout;
+export default SigninLayout;
